requirements.txt
Mastodon.py
pyTelegramBotAPI
requests
